Update and rename test1.py to test_compute_plot_phase.py
Adds bandpass filtering and detection of movement epochs.
Updates example plot
